When disconnected from the net, make sure we free all HTTP clients/servers             Completely remove IRQs from CC3k code             Speed up CC3k code, try and make it automatically power cycle it on timeouts

+++ targets/stm32/jshardware.c
@@ -674,7 +674,7 @@ void jshInterruptOn() {
 //int JSH_DELAY_OVERHEAD = 0;
 int JSH_DELAY_MULTIPLIER = 1;
 void jshDelayMicroseconds(int microsec) {
-  int iter = (microsec * JSH_DELAY_MULTIPLIER) >> 10;
+  int iter = (int)(((long long)microsec * (long long)JSH_DELAY_MULTIPLIER) >> 10);
 //  iter -= JSH_DELAY_OVERHEAD;
   if (iter<0) iter=0;
   while (iter--) __NOP();
